+++ src/index.ts
@@ -1,1 +1,2 @@
 export {default as QuadTree} from "./QuadTree";
+export type {Quadrants, BoundedItem} from "./QuadTree";